Refactor usage of cobra
